double changed to decimal

+++ ViewModels/ProductCreateViewModel.cs
@@ -4,6 +4,6 @@ public class ProductCreateViewModel
 {
     public string Name { get; set; }
     public string Description { get; set; }
-    public double BasePrice { get; set; }
-    public double TaxRate { get; set; }
+    public decimal BasePrice { get; set; }
+    public decimal TaxRate { get; set; }
 }
